+++ gems/aws-sdk-cloudwatch/lib/aws-sdk-cloudwatch/errors.rb
@@ -52,7 +52,6 @@ module Aws::CloudWatch
   #
   # Additionally, error classes are dynamically generated for service errors based on the error code
   # if they are not defined above.
-  # Some existing error classes may use a different class name than the one documented.
   module Errors
 
     extend Aws::Errors::DynamicErrors