+++ gems/aws-sdk-core/lib/aws-sdk-core/xml/error_handler.rb
@@ -5,6 +5,7 @@ require "cgi/util" if RUBY_VERSION < "3.5"
 
 module Aws
   module Xml
+    # @api private
     class ErrorHandler < Aws::ErrorHandler
 
       def call(context)