+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Document incorrect behavior in protocol error parsing (specifically around query and query compatible services).
+
 3.226.1 (2025-06-24)
 ------------------